Python Step 122 Drill
Drill 2 of a 2 part drill in the Tech Academy Python module with the following drill description:

Drill Description:
For this drill, you will need to write a script that creates a GUI with a button widget and a text widget. Your script will also include a function that when it is called will invoke a dialog modal which will allow users with the ability to select a folder directory from their system. Finally, your script will show the user’s selected directory path into the text field.